Use unittest discover automation.
