+++ python/lib/requirements.txt
@@ -1,0 +1,1 @@
+pyproj==3.4.0
